refactor: play 도메인의 컨트롤러, 서비스 클래스 이름 수정
- 이름을 통일화하고, 조금 더 도메인에 맞게 이름 수정
- GamePlayEngine 은 presentation 단에서 요청을 받도록 패키지 이동
