+++ src/main/kotlin/yjh/cstar/game/application/GameService.kt
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import yjh.cstar.game.domain.GameStartCommand
 import yjh.cstar.member.application.MemberService
-import yjh.cstar.play.application.GamePlayService
 import yjh.cstar.play.application.request.QuizDto
+import yjh.cstar.play.presentation.GamePlayEngine
 import yjh.cstar.quiz.application.QuizService
 import yjh.cstar.room.application.RoomService
 
@@ -15,7 +15,7 @@ class GameService(
     val memberService: MemberService,
     val roomService: RoomService,
     val quizService: QuizService,
-    val gamePlayService: GamePlayService,
+    val gamePlayEngine: GamePlayEngine,
 ) {
     @Transactional
     fun start(command: GameStartCommand) {
@@ -25,7 +25,7 @@ class GameService(
 
         val randomQuizData: List<QuizDto> = getRandomQuizData(command)
 
-        gamePlayService.start(playerInfo, randomQuizData, command.roomId, command.quizCategoryId)
+        gamePlayEngine.start(playerInfo, randomQuizData, command.roomId, command.quizCategoryId)
     }
 
     private fun getRandomQuizData(command: GameStartCommand) =
