+++ src/main/kotlin/yjh/cstar/play/application/GamePlayService.kt
@@ -1,21 +1,31 @@
 package yjh.cstar.play.application
 
-import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
+import yjh.cstar.game.application.GameResultService
+import yjh.cstar.play.application.port.AnswerProvider
+import yjh.cstar.play.application.port.GameNotifier
+import yjh.cstar.play.application.port.RankingHandler
 import yjh.cstar.play.application.request.QuizDto
-import yjh.cstar.util.Logger
+import yjh.cstar.play.application.request.toModel
+import yjh.cstar.play.domain.QuizGame
+import yjh.cstar.play.domain.game.GameInfo
+import yjh.cstar.room.application.RoomService
 
 @Service
 class GamePlayService(
-    private val quizGameService: QuizGameService,
+    private val answerProvider: AnswerProvider,
+    private val gameNotifier: GameNotifier,
+    private val rankingHandler: RankingHandler,
+    private val gameResultService: GameResultService,
+    private val roomService: RoomService,
 ) {
 
-    @Async("GamePlayThreadPool")
-    fun start(players: Map<Long, String>, quizzes: List<QuizDto>, roomId: Long, categoryId: Long) {
-        Logger.info("[INFO] 게임 엔진 스레드 시작 - roomId : $roomId")
+    fun play(players: Map<Long, String>, randomQuizzes: List<QuizDto>, roomId: Long, categoryId: Long) {
+        val gameInfo = GameInfo.of(players, randomQuizzes.map { it.toModel() }, roomId, categoryId)
+        val quizGame = QuizGame(gameInfo, answerProvider, gameNotifier, rankingHandler, gameResultService, roomService)
 
-        quizGameService.play(players, quizzes, roomId, categoryId)
-
-        Logger.info("[INFO] 게임 엔진 스레드 종료 - roomId : $roomId")
+        quizGame.initialize()
+        quizGame.run()
+        quizGame.finishGame()
     }
 }
